fix: address PR #10 round-2 review feedback
1. Rust smallint/i32 type mismatch: add ::int cast to survey_year
   in pg_trend_repository and pg_tls_repository SQL queries to
   prevent sqlx INT2→i32 decode errors at runtime

2. Pref-scoped MAX(survey_year): stats land_price query now uses
   MAX(survey_year) with the same pref_code filter so different
   prefectures with different latest years get correct results

3. L01 raw_pattern: add {pref_code} to prevent matching all
   prefectures' archives when converting for a single prefecture

4. KSJ field fallbacks in import_db.py: add _prop() helper that
   tries canonical names then KSJ-coded names (L01_006, A29_004,
   P29_004, P04_002, S12_001, N02_003, etc.)

5. admin_boundaries city_code: derive from adminCode length (5-digit
   JIS code = municipality) instead of relying on non-existent
   cityCode field

+++ services/backend/src/infra/pg_tls_repository.rs
@@ -109,7 +109,7 @@ impl TlsRepository for PgTlsRepository {
                 ORDER BY dist
                 LIMIT 1
             )
-            SELECT lp.survey_year AS year, lp.price_per_sqm, lp.address,
+            SELECT lp.survey_year::int AS year, lp.price_per_sqm, lp.address,
                    ST_Distance(lp.geom::geography, ST_SetSRID(ST_MakePoint($1, $2), 4326)::geography) AS distance_m
             FROM land_prices lp
             INNER JOIN nearest n ON lp.address = n.address
